Create push_delete_pair
This code re-does the push function, in addition to the delete functions and pair functionality. 
In the code, the delete function is passed a momento name and a momento "id," which is just the key that is associated with the momento value in the database. 

The inviteUserToMomento is the "pairing" feature. The user will provide an email address, and a momento name, and the function the associates that email with a momento. There are two helper functions for that function. These should be relatively simple to implement in the code.